+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.5.0] - Unreleased
+
+### Changed
+
+- If file is not specified on Console then the Console.file will return the current sys.stdout
+
+### Added
+
+- Added stderr parameter to Console
+- Added rich.reconfigure
+
 ## [9.4.0] - 2020-12-12
 
 ### Added